+++ CHANGELOG.md
@@ -5,6 +5,14 @@ Version changes are pinned to SDK releases.
 
 ## Unreleased
 
+## [1.4.2] 2023-07-05
+
+- Add in per asset risk state for cross margin accounts. ([#247](https://github.com/zetamarkets/sdk/pull/247))
+
+## [1.3.1] 2023-06-30
+
+- Support both marginAccount and crossMarginAccount in forceCancelOrderByOrderId() and forceCancelOrders(). ([#245](https://github.com/zetamarkets/sdk/pull/245))
+
 ## [1.2.0] 2023-06-29
 
 - Improve CrossMarginAccount migration UX. ([#244](https://github.com/zetamarkets/sdk/pull/244))